requirements.txt
anyio==3.6.1
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
fastapi==0.79.0
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.3
mysql-connector-python==8.0.30
packaging==21.3
pip==22.1.2
protobuf==3.20.1
pydantic==1.9.1
pyparsing==3.0.9
python-dotenv==0.20.0
PyYAML==6.0
redis==4.3.4
requests==2.28.1
setuptools==61.2.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
wheel==0.37.1
wrapt==1.14.1
pyodbc